requirements.txt
django==4.1.3
djangorestframework==3.14.0
dj-rest-auth==2.2.5
psycopg2==2.9.5
dj-rest-auth[with_social]
django-cors-headers